Implemented collisions and lives
I implemented a primitive version of collisions between the ship and
other game objects as well as added the 3 lives. Currently, when the
player loses a life, the console writes how many lives remain and when
the number of lives hits 0, the ship is destroyed. From here, we need to
implement the UI to reflect the number of lives remaining, implement
sounds/animations for when a collision happens, and the game over screen
for when the number of lives hits 0.